requirements.txt
openai>=1.0.0
asyncio
dataclasses
typing-extensions>=4.0.0
python-dotenv